:recycle: refactor

+++ app/api/prompt/[id]/route.ts
@@ -1,6 +1,8 @@
 import { NextRequest } from "next/server";
 import prisma from "@utils/auth/db/client";
 import { serverService } from "@utils/api-service";
+import { getServerSession } from "next-auth";
+import { authOptions } from "@utils/auth/authOptions";
 
 //! GET (read)
 export const GET = async (
@@ -25,6 +27,22 @@ export const PATCH = async (
   req: NextRequest,
   { params }: { params: { id: string } }
 ) => {
+  const session = await getServerSession(authOptions);
+
+  if (!session?.user) {
+    return new Response(
+      JSON.stringify({
+        error: {
+          message: "Unauthorized",
+          status: 401,
+        },
+      }),
+      {
+        status: 401,
+      }
+    );
+  }
+
   const { text, tag } = (await req.json()) as {
     text: string;
     tag: string;
@@ -52,6 +70,22 @@ export const DELETE = async (
   req: NextRequest,
   { params }: { params: { id: string } }
 ) => {
+  const session = await getServerSession(authOptions);
+
+  if (!session?.user) {
+    return new Response(
+      JSON.stringify({
+        error: {
+          message: "Unauthorized",
+          status: 401,
+        },
+      }),
+      {
+        status: 401,
+      }
+    );
+  }
+
   try {
     const {} = await prisma.prompt.delete({
       where: {
